treewide: do not use constructors to register codecs
Instead of using `__attribute__((constructor))` to create
functions which register the codecs, use `__attribute__((section))`
to put all NVCodec structs into a particularly named section of the
file (currently "nvd_codecs").

After this change, `NVCodecHolder` becomes unnecessary, so it
is removed; and no dynamic allocations are needed for codec
registration. Codecs can be directly iterated over, as if
they were defined in an array.