dynamic loader: Introduce use of nv-codec-headers
nv-codec-headers provides a dynamic symbol loader that makes it much
easier to build cuda/nvdec/nvenc dependent code and it's used by ffmpeg
and mpv. This driver makes heavy use of cuda and nvdec, and so it too
can benefit from having a dynamic loader to avoid requiring the cuda
SDK at build time.

I had to introduce the EGLStream related symbols and data types into
nv-codec-headers, but with those in place, it's a simple mechanical
process to switch over.

Fixes #1.