Implemented an awful hack to attempt to get VP8 working.
For VP8 NVDEC wants the full frame slice, but VA-API only gives us part of it, it's missing the first
few bytes. Luckily ffmpeg passes us a pointer to the original buffer when call vaCreateBuffer. We use
that to detect when a VP8 SliceData buffer is being created, and modify the incoming pointer to capture
the whole slice.

This hack relies on it being ffmpeg calling us, and is quite likely to break if something else calls us.