Check for modeset=1 before initialising.
This driver requires this parameter be set to function correctly, so we should attempt to check that it's actually
set, and fail appropriately if it isn't.

Checking this parameter is complicated by the fact we aren't root, and can't read the parameter from /sys, so we have
to use an ioctl to the matching DRM device file for the EGLDevice.