Reworked surface/pictureId allocation.
Due to Chrome not supplying the render target list upfront, we need to assign pictureIds else where. In addition we need some way to calculate the number of
decode surfaces needed. The chosen solution is to assign an id to the surface when allocates, and use the amount of allocated surfaces as the number of decode
surfaces.