Try to use the render node, instead of the main drm device node.
Opening a DRM master node attempts to make you the master of it, which in the case of the NVIDIA drivers, logs an error
if there's an existing master. We don't actually need to be a master, and the same caps are available through the render
node, so we should try to use that if it exists. It might not be available on older drivers, so we should still fallback
to the device node.

Fixes #41