Implement multithreaded surface resolve.
When decoding a frame, there is a need to map that frame to copy its
contents to the output buffer. However this mapping is a blocking call that
waits on the GPU to finish processing it. This change moves that call to a
separate thread to avoid blocking the calling process.