Added NVD_GPU environment variable.
In systems with multiple GPUs the user may not be using device 0, so we need a way for the user to specify which CUDA device we use.
This commit adds the NVD_GPU environment variable, which controls that.

The ideal solution is to move to using the device VA-API supplies us with, but that would likely cause more problems while we still
need LIBVA_DRIVER_NAME set.