Reworked GPU selection.
Previously we would select a specific GPU id (0 by default, unless overridden by NVD_GPU) to use at start up, ignoring the DRM FD
that VA-API passes in for us.

This commit changes that so we go looking for the EGLDevice that matches the fd that was passed in (by using part of the
minor device id), or if NVD_GPU is set we look for a specific GPU id.

This change should also allow us to detect optimus setups and potentially support them in the future.

Related #60