vabackend: make NVCodec struct aligned
Add the `__attribute__((aligned))` attribute to the `NVCodec`
type so that it has "the maximum alignment for the target".
This is needed because these structs are placed into a
section of the executable and the linker will align
the objects to e.g. 16-byte boundaries on x86-64 regardless
of the actual size of the object.

Currently, the size of this struct is just right, so this
is technically not needed, but should it be extended in the
future, this will prevent surprises.