vabackend: support 10/12 bit formats
The latest nvidia 520 drivers include support for the required DRM
formats to allow exporting of 10/12 bit surfaces.

The driver currently has incomplete support and various bits and
pieces commented out.

This change is an initial attempt to implement the required support for
the relevant profiles.

Things are simple where the profile and bit depth have a 1:1
correspondence, but AV1Profile0 includes both 8 and 10 bit content,
while VP9Profile2 includes 10 and 12 bit.

In these cases, we have to have a way to learn from the caller what
format is desired. In the most straight-forward case, the call to
createConfig will specify which format the caller wants as an RtFormat,
but this is optional. They are allowed to specify nothing.

On the other hand, other callers can provide pre-allocated surfaces
when calling createContext, and we can look at the surface format to
work out what they want.

What if a caller didn't provide either hint? We'd end up having to
guess and just assume 8bit. In practice, we believe that all relevant
callers use one of the above patterns.