Use NVD_GPU to control which DRM node to open in direct backend mode
When running in direct backend mode, we need to open the DRM node
ourselves, and the first node may not be the right one.

We will iterate through the DRM nodes looking for NVIDIA GPUs and will
use the first one found or the one identified by NVD_GPU if set.