direct: error out when drmFd == -1
Currently we try really hard to guess which is the correct fd, when the
frontend (wayland) or client (when using drm) hasn't given us one...
aka when X11 is used.

Despite the best of intent, we shouldn't be doing that since it uses a
device which is not associated with the X display. Perhaps one day the
Nvidia driver will support DRI2/DRI3, which will let the libva give us a
meaningful fd.

This has the side-effect of temporary disabling NVD_GPU override.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>