vabackend: Treat surfaceCount == 1 the same as surfaceCount == 0
When Chrome(ium) wants dynamic surface allocation it starts with an initial
count of 0, and that is handled today. On the other hand, recent versions of
ffmpeg have switched to preferring dynamic allocation but they start with a
surfaceCount of 1. If we treat this as a legitimate static request, we cannot
do any decoding in practice, as you always need more than 1 surface.

So, treat a surfaceCount of 1 the same as 0, and use our same dynamic
allocation work-around.