Use uintptr_t for pointers-to-interegs casts.
On 32-bit targets, casting pointers directly to uint64_t and similar
64-bit typedefs causes warnings with gcc. The result of such casts is
not defined if the most significant bit of the pointer representation
is set (i.e. the resulting integer may get sign-extended or zero-extended).

Use uintptr_t for the intermediate cast to unsigned integer which is
then converted to the resulting 64-bit integer (by zero-extending).