Compute and cache supported codecs.
Since Firefox has moved to a single driver instance this has caused a lot of calls to the nvidia driver to determine codec support on every new context created. This has started to cause problems as there seems to be a bug in some driver versions that can crash the driver if a destroy context call coincides with one of the support checks.