vp8: validate vp8 header and insert null if missing
Chromium's VA implementation passes slices starting at the frame data,
i.e. without the header, while Firefox passes the header correctly.

The previous commit caused the real header to be offset too, causing
corruption in firefox. This commit validates the header on the first keyframe
of the video, and only inserts null headers to the data if the first keyframe
has no header.